+++ build_neighborhood.h
@@ -1,5 +1,5 @@
 #include "image.h"
 
-double *build_neighborhood(Image *G, int L, int x, int y);
+int *build_neighborhood(Image *G, int L, int x, int y);
 
 int match_neighborhood(double *Na, double *Ns, int l);
